+++ Lib/fontbakery/profiles/googlefonts.py
@@ -43,6 +43,7 @@ OVERRIDDEN_CHECKS = [
     "com.google.fonts/check/italic_axis_in_stat_is_boolean",
     "com.google.fonts/check/italic_axis_last",
     "com.google.fonts/check/alt_caron",
+    "com.google.fonts/check/legacy_accents",
 ]
 
 METADATA_CHECKS = [
@@ -7405,6 +7406,13 @@ profile.check_log_override(
     "com.google.fonts/check/alt_caron",
     overrides=(("bad-mark", FAIL, KEEP_ORIGINAL_MESSAGE),),
     reason=("For Google Fonts, one of the comma-lookalikes is a FAIL"),
+)
+
+profile.check_log_override(
+    # From universal.py
+    "com.google.fonts/check/legacy_accents",
+    overrides=(("legacy-accents-component", FAIL, KEEP_ORIGINAL_MESSAGE),),
+    reason=("For Google Fonts, using legacy accents as components is a FAIL."),
 )
 
 GOOGLEFONTS_PROFILE_CHECKS = add_check_overrides(